* the `ClickListenerHelper` was renamed to `EventHookUtil` and moved to the utils package. It is now a util as it no longer needs an instance and has just 3 public static methods. * `withItemEvent` is now deprecated in favor of the new naming `withEventHook` * simplify and remove code
