+++ app/src/main/java/com/mikepenz/fastadapter/app/ImageListActivity.java
@@ -13,17 +13,14 @@ import android.widget.Toast;
 
 import com.mikepenz.fastadapter.FastAdapter;
 import com.mikepenz.fastadapter.IAdapter;
-import com.mikepenz.fastadapter.commons.adapters.FastItemAdapter;
 import com.mikepenz.fastadapter.app.dummy.ImageDummyData;
 import com.mikepenz.fastadapter.app.items.ImageItem;
-import com.mikepenz.fastadapter.helpers.ClickListenerHelper;
+import com.mikepenz.fastadapter.commons.adapters.FastItemAdapter;
 import com.mikepenz.materialize.MaterializeBuilder;
 
 public class ImageListActivity extends AppCompatActivity {
     //save our FastAdapter
     private FastItemAdapter<ImageItem> mFastItemAdapter;
-
-    private ClickListenerHelper<ImageItem> mClickListenerHelper;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -41,9 +38,6 @@ public class ImageListActivity extends AppCompatActivity {
 
         //create our FastAdapter which will manage everything
         mFastItemAdapter = new FastItemAdapter<>();
-
-        //init the ClickListenerHelper which simplifies custom click listeners on views of the Adapter
-        mClickListenerHelper = new ClickListenerHelper<>();
 
         //configure our fastAdapter
         mFastItemAdapter.withOnClickListener(new FastAdapter.OnClickListener<ImageItem>() {
@@ -77,7 +71,7 @@ public class ImageListActivity extends AppCompatActivity {
         //a custom OnCreateViewHolder listener class which is used to create the viewHolders
         //we define the listener for the imageLovedContainer here for better performance
         //you can also define the listener within the items bindView method but performance is better if you do it like this
-        mFastItemAdapter.withItemEvent(new ImageItem.ImageItemHeartClickEvent());
+        mFastItemAdapter.withEventHook(new ImageItem.ImageItemHeartClickEvent());
 
         //set the back arrow in the toolbar
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
